automations/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
supabase==2.7.0
pydantic==2.9.0
resend==2.3.0
python-multipart==0.0.9
httpx==0.27.0
python-jose==3.3.0
apscheduler==3.10.4
jinja2==3.1.4
python-dotenv==1.0.1
cryptography==43.0.1
aiosmtplib==3.0.1
email-validator==2.2.0
